oversight: Add global dirty flag
Instead of verifying the signal db and GetManagedObjects match anytime its
been 3+ seconds since last signal, we will only do so if a signal has
arrived.  This way we don't flag an error if the service is in the middle
of changing state before a signal is generated.  Technically, this could
be considered an error in the implementation of the service.

Signed-off-by: Tony Asleson <tasleson@redhat.com>